Added new Guide: Alternatives to CodeCademy
This is my first submission to the proposed articles. This guide has
been aproved by the moderators group as well. I am proposing it here
because I would like to know if I still need to change something in
order to ship this baby to the Help Center.

##Description:
Do you want to get to the next level and you think that CodeCademy is
not enough? Do you want to explore other ways of learning and other
resources? This article is just for you. Here we explore other websites
and forums that directly and indirectly compete with CodeCademy and we
divide our approach in three ways so that you can easily identify them.
Are you a competitive person, someone looking for a job or you just want
an official certificate? Some of our suggestions may interest you then,
wait no more and take a peek!