Fixed the problem !!
I removed the white space in the beginning of line 13 and removed all the tabs. Codecademy + Markdown + Tabs = disaster. I have also had this problem in my tutorials. If you really insist in indentation as to make the article easier to read, you may add a single white space before every number.

This is why I battle on an every day basis to ask CC employees to actually add a decent flavor of Markdown to CC xD